back/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
executing==2.0.1
fonttools==4.46.0
idna==3.4
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.7
mpldatacursor==0.7.1
nest-asyncio==1.6.0
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==26.0.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.13
